Accept raw definitions in container factories
S.array, S.record, S.optional, S.nullable, S.nullish, S.compactColumns
(and S.list) now accept anything S.schema accepts — literals, plain
objects/arrays with nested schemas — and infer the schema from it, the
way S.union and S.schema already do: S.array({ id: S.string }) is
S.array(S.schema({ id: S.string })).

Runtime: definitionToSchema/traverseDefinition move from factory.ts down
to composites.ts (all their deps already live at or below it), so the
factories can convert their argument. Each converting factory is a thin
public wrapper over an Internal-taking one (arrayFactory/dictFactory);
internal callers use those directly — both to skip the no-op conversion
on codegen-hot paths and because the S.json singleton builds its members
during module init, before operations.ts installs the `~standard`
prototype marker that definitionToSchema detects schemas by.

Types: the factories take `SchemaLike<TInput, TOutput> | TDef` in a
single signature — a schema resolves through the structural constituent
without touching the recursive UnknownTo* machinery; only a raw
definition falls into the naked TDef param and pays for the defaults.
A SchemaLike/definition overload pair measured much cheaper for schema
args but silently breaks `schema.with(S.optional, …)`: TS only infers
through a function value with exactly one call signature.

Metrics: full-bundle size is flat (+1 B). Per-export bundles that newly
retain the definition machinery grow: optional/nullable/nullish +2.7 KB
(they previously shipped no object/array decoders), array/dict/record/
compactColumns/list ~+0.2 KB. ts.instantiations pay ~+42 per factory
call (e.g. record 392→434, nullable 380→422; optional-object worst at
2754→3116). Both are the cost of the feature, minimized as above.
