+++ packages/sury/specs/array-length.yaml
@@ -7,7 +7,7 @@ ts:
   schema: S.array(S.string).with(S.length, 2)
   input: string[]
   output: string[]
-  instantiations: 5290
+  instantiations: 5332
 vs:
   zod: z.array(z.string()).length(2)
 jsonSchema:
