+++ packages/sury/specs/bundleSize.yaml
@@ -1,106 +1,106 @@
 # Minified+gzipped bytes per public export of index.mjs, plus `total` for the whole entry.
 # Generated by `pnpm spec check --write` — every row is measured, so never hand-write one.
-total: 24325
+total: 24326
 exports:
-  $res_Metadata_Id_make: 4135
-  $res_Metadata_get: 4129
-  $res_Metadata_set: 4147
-  $res_Option_getOr: 9753
-  $res_Option_getOrWith: 9753
-  $res_assertAsyncOrThrow: 4470
-  $res_nullAsOption: 12030
-  $res_nullAsUnit: 4485
-  $res_nullableAsOption: 9394
-  $res_option: 12005
-  $res_pathConcat: 4127
-  $res_pathFromArray: 4217
-  $res_pathFromLocation: 4173
-  $res_pathToArray: 4127
-  $res_schema: 12222
-  $res_setExnId: 4130
-  $res_transform: 4610
-  $res_unit: 4450
-  Error: 4127
-  any: 4120
-  array: 12001
+  $res_Metadata_Id_make: 4134
+  $res_Metadata_get: 4128
+  $res_Metadata_set: 4145
+  $res_Option_getOr: 9751
+  $res_Option_getOrWith: 9751
+  $res_assertAsyncOrThrow: 4471
+  $res_nullAsOption: 12023
+  $res_nullAsUnit: 4483
+  $res_nullableAsOption: 9391
+  $res_option: 11999
+  $res_pathConcat: 4126
+  $res_pathFromArray: 4215
+  $res_pathFromLocation: 4171
+  $res_pathToArray: 4126
+  $res_schema: 12215
+  $res_setExnId: 4128
+  $res_transform: 4609
+  $res_unit: 4451
+  Error: 4126
+  any: 4119
+  array: 12208
   assert: 4502
-  asyncDecoder: 4134
-  asyncDecoderAssert: 4643
-  asyncEncoder: 4157
-  asyncParser: 4135
-  bigint: 4423
-  bool: 4415
-  boolean: 4415
-  brand: 4136
-  compactColumns: 12952
-  cuid: 4525
-  date: 4455
-  decoder: 4127
-  deepStrict: 4239
-  deepStrip: 4236
-  dict: 12004
-  email: 4571
-  empty: 4827
-  enableStandardJSONSchema: 5578
-  encoder: 4154
-  enum: 12226
+  asyncDecoder: 4132
+  asyncDecoderAssert: 4642
+  asyncEncoder: 4155
+  asyncParser: 4133
+  bigint: 4421
+  bool: 4413
+  boolean: 4413
+  brand: 4135
+  compactColumns: 13166
+  cuid: 4524
+  date: 4453
+  decoder: 4126
+  deepStrict: 4237
+  deepStrip: 4235
+  dict: 12208
+  email: 4570
+  empty: 4826
+  enableStandardJSONSchema: 5576
+  encoder: 4152
+  enum: 12216
   extendJSONSchema: 4211
-  float: 4545
-  fromJSONSchema: 18831
-  global: 4173
-  gt: 4836
-  gte: 4834
-  inputExpression: 4119
-  instance: 4298
-  int: 4569
-  int32: 4569
-  is: 4518
-  isAsync: 4192
-  isoDateTime: 4562
-  json: 13431
-  jsonString: 13876
-  jsonStringWithSpace: 13892
-  length: 4817
-  list: 12519
-  literal: 12213
-  lt: 4838
-  lte: 4837
+  float: 4543
+  fromJSONSchema: 18848
+  global: 4172
+  gt: 4835
+  gte: 4833
+  inputExpression: 4118
+  instance: 4296
+  int: 4567
+  int32: 4567
+  is: 4517
+  isAsync: 4190
+  isoDateTime: 4560
+  json: 13428
+  jsonString: 13870
+  jsonStringWithSpace: 13885
+  length: 4816
+  list: 12720
+  literal: 12204
+  lt: 4837
+  lte: 4836
   maxLength: 4814
-  merge: 12129
-  meta: 4265
+  merge: 12122
+  meta: 4263
   minLength: 4814
   nan: 4455
-  never: 4187
-  noValidation: 4136
-  nonEmpty: 4824
-  nullable: 9828
-  nullish: 9368
-  number: 4545
-  object: 13947
-  optional: 9802
-  outputExpression: 4130
-  parser: 4133
-  pattern: 4328
-  port: 4646
-  record: 12004
-  recursive: 4824
-  refine: 4338
-  reverse: 4121
-  safe: 4166
-  safeAsync: 4186
-  schema: 12213
-  shape: 13149
-  strict: 4238
-  string: 4411
-  strip: 4236
-  symbol: 4327
-  to: 4387
-  toJSONSchema: 5549
-  trim: 4637
-  tuple: 13271
-  uint8Array: 4582
-  union: 12218
-  unknown: 4120
-  url: 4524
-  uuid: 4542
+  never: 4185
+  noValidation: 4135
+  nonEmpty: 4823
+  nullable: 12543
+  nullish: 12215
+  number: 4543
+  object: 13943
+  optional: 12513
+  outputExpression: 4128
+  parser: 4131
+  pattern: 4326
+  port: 4644
+  record: 12208
+  recursive: 4825
+  refine: 4335
+  reverse: 4120
+  safe: 4164
+  safeAsync: 4184
+  schema: 12204
+  shape: 13140
+  strict: 4236
+  string: 4408
+  strip: 4234
+  symbol: 4325
+  to: 4386
+  toJSONSchema: 5548
+  trim: 4635
+  tuple: 13266
+  uint8Array: 4583
+  union: 12208
+  unknown: 4119
+  url: 4522
+  uuid: 4540
   void: 4458
