+++ packages/sury/specs/codec-json-optional-union-literal-first.yaml
@@ -3,7 +3,7 @@ ts:
   schema: 'S.object({ selector: S.optional(S.union([S.literal("*"), S.string, S.array(S.string)])) }).with(S.to, S.json)'
   input: "{ selector?: string | string[] | undefined; }"
   output: JSON
-  instantiations: 8075
+  instantiations: 8157
 vs:
   zod:
     _skip: not-applicable
