+++ packages/sury/specs/codec-optional-literal-nullable-literal.yaml
@@ -3,7 +3,7 @@ ts:
   schema: S.optional(S.literal("x")).with(S.to, S.nullable(S.literal("x")))
   input: '"x" | undefined'
   output: '"x" | null'
-  instantiations: 5677
+  instantiations: 5768
 vs:
   zod:
     _skip: not-applicable
