+++ packages/sury/specs/codec-optional-nullable-partial.yaml
@@ -3,7 +3,7 @@ ts:
   schema: S.optional(S.string).with(S.to, S.nullable(S.boolean))
   input: string | undefined
   output: boolean | null
-  instantiations: 5376
+  instantiations: 5458
 vs:
   zod:
     _skip: not-applicable
