+++ packages/sury/specs/codec-optional-nullable-transformed.yaml
@@ -3,7 +3,7 @@ ts:
   schema: S.optional(S.string).with(S.to, S.nullable(S.string.with(S.to, S.boolean)))
   input: string | undefined
   output: boolean | null
-  instantiations: 5743
+  instantiations: 5829
 vs:
   zod:
     _skip: not-applicable
