+++ packages/sury/specs/codec-optional-nullable.yaml
@@ -3,7 +3,7 @@ ts:
   schema: S.optional(S.string).with(S.to, S.nullable(S.string))
   input: string | undefined
   output: string | null
-  instantiations: 5313
+  instantiations: 5400
 vs:
   zod:
     _skip: not-applicable
