+++ packages/sury/specs/codec-string-optional-partial.yaml
@@ -3,7 +3,7 @@ ts:
   schema: S.string.with(S.to, S.optional(S.string))
   input: string
   output: string | undefined
-  instantiations: 5229
+  instantiations: 5271
 vs:
   zod:
     _skip: not-applicable
