+++ packages/sury/specs/compact-columns.yaml
@@ -3,7 +3,7 @@ ts:
   schema: "S.compactColumns(S.unknown).with(S.to, S.array(S.schema({ id: S.number, name: S.string })))"
   input: unknown[][]
   output: "{ id: number; name: string; }[]"
-  instantiations: 6676
+  instantiations: 6887
 vs:
   zod:
     _skip: not-applicable
