+++ packages/sury/specs/object-in-object3.yaml
@@ -6,7 +6,7 @@ ts:
     - "S.object({a: S.string, nested: {b: S.number, inner: {c: S.boolean, d: S.optional(S.string)}}})"
   input: "{ a: string; nested: { b: number; inner: { c: boolean; d?: string | undefined; }; }; }"
   output: "{ a: string; nested: { b: number; inner: { c: boolean; d?: string | undefined; }; }; }"
-  instantiations: 2769
+  instantiations: 2809
 vs:
   zod: "z.object({ a: z.string(), nested: z.object({ b: z.number(), inner: z.object({ c: z.boolean(), d: z.string().optional() }) }) })"
 jsonSchema:
