+++ packages/sury/specs/object5-optional.yaml
@@ -5,7 +5,7 @@ ts:
     - "S.object({a: S.string, b: S.optional(S.number), c: S.boolean, d: S.optional(S.bigint), e: S.symbol})"
   input: "{ a: string; c: boolean; e: symbol; b?: number | undefined; d?: bigint | undefined; }"
   output: "{ a: string; c: boolean; e: symbol; b?: number | undefined; d?: bigint | undefined; }"
-  instantiations: 3610
+  instantiations: 3676
 vs:
   zod: "z.object({ a: z.string(), b: z.number().optional(), c: z.boolean(), d: z.bigint().optional(), e: z.symbol() })"
 jsonSchema:
