+++ packages/sury/specs/optional-definition.yaml
@@ -1,0 +1,29 @@
+# yaml-language-server: $schema=./spec.schema.json
+ts:
+  schema: "S.optional({ a: S.boolean })"
+  input: "{ a: boolean; } | undefined"
+  output: "{ a: boolean; } | undefined"
+  instantiations: 1354
+vs:
+  zod: "z.object({ a: z.boolean() }).optional()"
+jsonSchema:
+  input: "Expected JSON, received { a: boolean; } | undefined"
+  output: "Expected JSON, received { a: boolean; } | undefined"
+operations:
+  parse:
+    expression: i=>{for(;;){if(typeof i==="object"&&i&&!Array.isArray(i)){let v0=i["a"];typeof v0==="boolean"||e[0](v0);i={"a":v0,};break}if(i===void 0)break;e[1](i)}return i}
+    examples:
+      object:
+        input: "{ a: true }"
+        output: "{ a: true }"
+      undefined:
+        input: undefined
+        output: undefined
+      wrong-member-type:
+        input: "{ a: 1 }"
+        error: 'Failed at ["a"]: Expected boolean, received 1'
+      null-is-not-an-object:
+        input: "null"
+        error: "Expected { a: boolean; } | undefined, received null"
+  decode: identity
+  encode: identity
