+++ packages/sury/specs/optional-object.yaml
@@ -3,7 +3,7 @@ ts:
   schema: "S.optional(S.schema({ a: S.optional(S.string) }))"
   input: "{ a?: string | undefined; } | undefined"
   output: "{ a?: string | undefined; } | undefined"
-  instantiations: 2754
+  instantiations: 3116
 vs:
   zod: "z.object({ a: z.string().optional() }).optional()"
 jsonSchema:
