+++ packages/sury/src/advanced/json.ts
@@ -50,8 +50,8 @@ import {
   B_varWithoutAllocation,
 } from "../builder";
 import {
-  array,
   arrayDecoder,
+  arrayFactory,
   completeObjectVal,
   dictFactory,
   makeObjectVal,
@@ -87,7 +87,7 @@ export const jsonEncoderFn = (input: Val, target: Internal): Val => {
   } else if (flagUnsafeHas(toTagFlag, tagFlagArray)) {
     // Validate that the input is an array
     // and then update the schema to be an array of json instead of array of unknown
-    const jsonExpected = array(unknown);
+    const jsonExpected = arrayFactory(unknown);
     const output = parse(B_refine(input, unknown, U, jsonExpected));
     output.s.additionalItems = json;
     output.e = target;
@@ -254,7 +254,7 @@ export const json: Internal = /* @__PURE__ */ initSchema(refTag, (s) => {
     float,
     nullLiteral,
     dictFactory(jsonRef),
-    array(jsonRef),
+    arrayFactory(jsonRef),
   ];
   const has: Partial<Record<Tag, boolean>> = {};
   anyOf.forEach((schema) => {
