+++ packages/sury/src/advanced/list.ts
@@ -28,7 +28,7 @@ const listToArray = (list: RescriptList): unknown[] => {
 }
 
 // @__NO_SIDE_EFFECTS__
-export const list = (schema: Internal): Internal => {
+export const list = (schema: unknown): Internal => {
   return transform(array(schema), () => ({
     p: (array: unknown) => listFromArray(array as unknown[]),
     s: (list: unknown) => listToArray(list as RescriptList),
