+++ packages/sury/src/jsonschema.ts
@@ -43,8 +43,8 @@ import {
 } from "./base";
 import { json } from "./advanced/json";
 import { B_operationArg } from "./builder";
-import { array, option } from "./composites";
-import { definitionToSchema, schemaFactory } from "./factory";
+import { array, definitionToSchema, option } from "./composites";
+import { schemaFactory } from "./factory";
 import {
   meta,
   Metadata_get,
